+++ elements/pt-app.js
@@ -122,6 +122,12 @@ export default class PTAppElement extends HTMLElement {
                 </x-navitem>
               </a>
 
+              <a href="/elements/x-drawer">
+                <x-navitem>
+                  <x-label>x-drawer</x-label>
+                </x-navitem>
+              </a>
+
               <a href="/elements/x-icon">
                 <x-navitem>
                   <x-label>x-icon</x-label>